+++ src/utils/constants.py
@@ -1,0 +1,5 @@
+from pathlib import Path
+from typing import Final
+
+SAMPLES_PATH: Final[Path] = Path("./samples")
+EMBEDDING_SAMPLE_PATH = SAMPLES_PATH / "mock_embedding.json"