+++ CHANGELOG.md
@@ -11,5 +11,6 @@ All notable changes to this project will be documented in this file.
 - Created `GameReview` class in `src/models/gamereview.py`. [dev/aingeru]
 - Created `GameReviewEmbeddings` class in `src/models/gamereviewembeddings.py`. [dev/aingeru]
 - Created `EmbeddingService` class in `src/services/embedding_service.py`. [dev/aingeru]
+- `src/utils/constants.py`. [dev/lukas]
 
 ### Changed